+++ virtual-keyboard/src/template.html
@@ -4,7 +4,6 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="icon" type="image/png" sizes="32x32" href="favicon.ico">
     <title>RSS Virtual Keyboard</title>
 </head>
     <body>
